[#3007] Content Assistant test infrastructure improvements.
- Extend the Content Assistant test infrastructure to test proposals
from several resources: modify the AbstractContentAssistTest
getResourceFor method to ensure that the test file points to a file in
the test project.
- Uplift the testContentAssistant method from the project example
classes into the AbstractContentAssistTest and rename it to
assertContentAssistant.
- Add the events_from_another_file() test case to the
StatemachineContentAssistTest test cases to demonstrate how to test the
content assistant with several resources.
- Also add the ContentAssistWithSeveralResourcesTest test cases that are
continuously executed by the CI build.

Closes #3007

Signed-off-by: miklossy <miklossy@itemis.de>